+++ database/migrations/tenant/2022_05_13_163802_tenant_purchase_settlement_payments_table.php
@@ -13,18 +13,20 @@ class TenantPurchaseSettlementPaymentsTable extends Migration
      */
     public function up()
     {
-        Schema::create('purchase_settlement_payments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('purchase_settlement_id');
-            $table->date('date_of_payment');
-            $table->char('payment_method_type_id', 2);
-            $table->string('reference')->nullable();
-            $table->decimal('change',12, 2)->default(0);
-            $table->decimal('payment', 12, 2);
+        if (!Schema::hasTable('purchase_settlement_payments')) {
+            Schema::create('purchase_settlement_payments', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('purchase_settlement_id');
+                $table->date('date_of_payment');
+                $table->char('payment_method_type_id', 2);
+                $table->string('reference')->nullable();
+                $table->decimal('change', 12, 2)->default(0);
+                $table->decimal('payment', 12, 2);
 
-            $table->foreign('purchase_settlement_id')->references('id')->on('purchase_settlements')->onDelete('cascade');
-            $table->foreign('payment_method_type_id')->references('id')->on('payment_method_types');
-        });
+                $table->foreign('purchase_settlement_id')->references('id')->on('purchase_settlements')->onDelete('cascade');
+                $table->foreign('payment_method_type_id')->references('id')->on('payment_method_types');
+            });
+        }
     }
 
     /**
@@ -34,6 +36,8 @@ class TenantPurchaseSettlementPaymentsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('purchase_settlement_payments');
+        if (Schema::hasTable('purchase_settlement_payments')) {
+            Schema::dropIfExists('purchase_settlement_payments');
+        }
     }
 }
